requirements.txt
Flask
pytesseract
Pillow
openpyxl
Werkzeug
